+++ src/main/java/com/example/fantasticrestaurantv1/orders/repository/OrderRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Collection;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 
-    Collection<Order> findAllByUserId(Integer id);
+    Collection<Order> findAllByUserId(Long id);
 }
